Move verify_auth_key and AUTH_HEADER into auth.py
